Move folder to namespace
