+++ DSharpPlus/Entities/AuditLogs/Entries/DiscordAuditLogAutoModerationExecutedEntry.cs
@@ -23,7 +23,7 @@
 
 using DSharpPlus.Enums;
 
-namespace DSharpPlus.Entities;
+namespace DSharpPlus.Entities.AuditLogs;
 
 public sealed class DiscordAuditLogAutoModerationExecutedEntry : DiscordAuditLogEntry
 {
