+++ DSharpPlus/Entities/AuditLogs/Entries/DiscordAuditLogMemberUpdateEntry.cs
@@ -24,8 +24,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace DSharpPlus.Entities;
-
+namespace DSharpPlus.Entities.AuditLogs;
 public sealed class DiscordAuditLogMemberUpdateEntry : DiscordAuditLogEntry
 {
     /// <summary>
